% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y2print-package.R
\docType{package}
\name{y2print-package}
\alias{y2print}
\alias{y2print-package}
\title{y2print: Enhanced Printing for Frequency Tibbles}
\description{
Provides printing enhancments and QOL improvements for the display of frequency tibbles from the y2clerk package.
}
\author{
\strong{Maintainer}: Adam Hubbs \email{adam@y2analytics.com}

Authors:
\itemize{
  \item Adam Hubbs \email{adam@y2analytics.com}
}

}
\keyword{internal}
